68219.203904 uso - installWith [0.3.0](/scripts/diff/68219/203904)
- New [Rate and Limiting](/topics/52160) on USO makes detection slower... Now shows a grey Install button if not finished evaluating.  Users can still install while waiting however the DOM may change at any time within a few seconds of page load to allow for optional wrapping.
- rex tweaks

NOTES:

If a user is using [UserScript.org Install Link On All Script Pages](/scripts/show/55774) please ensure that it runs before this script in the GM Manage Dialog for proper execution.

Occasionally USO will now claim that a script is unlisted even though it's not... still reworking this since @uso:unlisted is a phantom key that wasn't approved last year and this glitch with Rate and Limiting change is hard to reproduce. :P :)

68219.203904

6c66f78e7e49f1f422bd283283a4b38bd392277a