69307.298780 uso - Count Issues [0.9.0](/scripts/diff/69307/298780)
- Removed migration from [0.4.0](/topics/46434?page=2#posts-273597) as people should already have the newest version by now
- ALPHA PRERELEASE of Show recent Versions on Source Code page
Please note this is not finished so it may be a little buggy and definitely feature incomplete. Toggle it on is in the [about page](/scripts/show/69307) to try it out.  This is not a guarantee that all element names will stay the same as I'm still refining the script.  As I near version 1.0 of Count Issues I will most likely be doing a complete rewrite of a few things to minimize the footprint but still allow maximum flexibility.

Users might also want to install [uso - Syntax Highlight](/scripts/show/85835) as this will help immensely with this new feature. :)

Recommended script order, assuming you have all of these, is something like this:
...[uso - screen.css](/scripts/show/48071)[userscripts.org alternate CSS](/scripts/show/34698)[uso - Count Issues](/scripts/show/69307)[uso - installWith](/scripts/show/68219)[uso - Syntax Highlight](/scripts/show/85835)...[uso - Anchor Bookmarks](/scripts/show/69725)...

69307.298780

995f67ef00d774b1e68d7cea897f2557b246242d