68219.300154 uso - installWith [0.9.0](/scripts/diff/68219/300154) and usoCheckup 1.0.198
-  NEW FEATURE  Introduction of Greasemonkey @icon support via the wrapper
Please note if a script already has an [@icon](http://sf.net/apps/mediawiki/greasemonkey/index.php?title=Metadata_Block#.40icon) installWith will NOT override it regardless of what you select.

There might be a false positive perhaps once in about 20 years or so when USO scriptids get up to 32 digits and the md5 sum matches it... but the probability of this is nearly nill. :)

To explain this feature a little bit in words... usoC will check to see if @icon is present... if it is... say bye bye to this feature.  If an end user selects the first icon, which should normally always be there, which is the gravatar... it will use that in Greasemonkey 0.9.x+ instead of the plain script icon... existing 0.8.x users, this feature is not useful so it is best to uncheck the related options.  All scripts from that same ScriptWright will use their personal gravatar for the GM @icon unless they have already added one to the script.

If an end user picks favor USO script icon... this is as it suggests... if and when screenshots and icon are ever reinstated this will pull from that icon field... if a ScriptWright hasn't done one usoC/installWith will fall back to using the gravatar of the author.

If everything fails... say amazonaws and gravatar both go out... usoC will do it's best to recover from this so it doesn't prevent installation of scripts from USO.  Current GM has an open ticket [#1214](http://github.com/greasemonkey/greasemonkey/issues/1214) to address this.  If you are on GitHub PLEASE GO VOTE IT UP.  In an extremely unlikely situation where the ScriptWright is changing the USO script icon and an end user is installing a script simultaneously... this may prevent script (re)installation... the "work around" is to wait a few seconds/minutes and try again till the ScriptWright is finished. Once a "failure" has been detected in automatic updating you are stuck with that icon unless you visit the script homepage and reinstall from there.

When the appropriate icons are available they will be loaded into the GM_config window reduced to the size that you will see them in Greasemonkeys Add-ons Manage Dialog... which is currently 32x32 pixels. When screenshots and/or icon is reenabled ideally a ScriptWright should upload a 1:1 ratio image (e.g. say 90x90 pixels) which will reduce cleaner than one with the 120x90 pixel 4:3 ratio.

If you encounter any issues please report them in the [ERRATA](/topics/45480) topic.

Test homepage(s)
[66548](/scripts/show/66548) This has a USO script icon that scales well from 4:3 to 1:1. Please note that script already has an updater... I still don't suggest double wrapping as this could cause browser hangs, updater malfunctions on both, etc. when the script is executed.[66531](/scripts/show/66531) This has a USO script icon that doesn't scale well from 4:3 to 1:1[67120](/scripts/show/67120) This does NOT have a USO script icon so it should use my gravatar instead.[92598](/scripts/show/92598) This has @icon in it and no updater, so the ScriptWrights choice will be used instead.[77895](/scripts/show/77895) LOL and has USO icon and stretches.

Enjoy! :)

68219.300154

10f2411430fe4051948fdfc9100ae32717a5cfdb