68219.301770 uso - installWith [0.9.3](/scripts/diff/68219/301770)
- Progress parseMeta routine to eliminate a false positive. Too much regression from [0.5.1](/topics/46434?page=2#posts-275419) method in Count Issues applied to installWith. This should still detect the true positive for those ScriptWrights trying to avoid detection. Confirmed by [this author](/users/257807) in [this script](/scripts/show/92059) with [this version](/scripts/version/92059/302137.user.js).

Example is in [this script](/scripts/show/87496) in [this version](/scripts/version/87496/301731.user.js) by [this author](/users/158918) ... shouldn't be returning with text before it.  Both installWith and Count Issues is currently non-discriminatory if a @key is even in the metadata block... This is by design to detect "combined scripts"... however those comments that start out with text and contain // @keys in that script shouldn't be returned. Combined scripts usually but not always are "botched" up sometimes with additional updaters and can cause DDoS so wrapping will not be immediately available if there is an issue detected.

68219.301770

351d39af586d0b45bef72ef75e32caba2c3ee3e9