69307.301771 uso - Count Issues [0.9.2](/scripts/diff/69307/301771)
- Progress parseMeta routine to eliminate a false positive. Too much regression from [0.5.1](/topics/46434?page=2#posts-275419).  This should still detect the true positive for those ScriptWrights trying to avoid detection. Confirmed by [this author](/users/257807) in [this script](/scripts/show/92059) with [this version](/scripts/version/92059/302137.user.js).

Example is in [this script](/scripts/show/87496) in [this diff](/scripts/version/87496/301731.user.js) by [this author](/users/158918) ... shouldn't be returning with text before it.  Both installWith and Count Issues is currently non-discriminatory if a @key is even in the metadata block... This is by design to detect "combined scripts"... however those comments that start out with text and contain // @keys in that script shouldn't be returned.

69307.301771

0a58bbc778b3c5c8bb041adebf54557c0779016b