68219.337533 uso - installWith [0.10.0](/scripts/diff/68219/337533)
- [This script](/scripts/show/103796) by [this author](/users/338966) contains embedded updater that accesses a [deleted script](/scripts/show/103455) on USO. Also present in [this script](/scripts/show/103494) by [this author](/users/132554).

Advisory: SEVERE

Please note this method won't catch this type of exploit 100% of the time so best to do a visual inspection on any script that has direct user.js references. (usually when it's greyed with "Install with" in the button)

Thought we were all past this issue with USO a while back but seems to have resurfaced. :(

68219.337533

7abf716174a6a87427d391ef4019fb05fdadb51f