69307.359673 uso - Count Issues [0.15.3](/scripts/diff/69307/359673)
- NEW FEATURE
Metadata block key names in the sidebar can now be clicked to my forked wiki which usually (hopefully) contains more information on what they are and do... as I've been told multiple times now.  If the @key doesn't exist from the user.js you are viewing then it will take you to the top of the page. These links are "silent" due to this possibility... I do not plan on adding every @key that I find... just the major ones that are helpful.  The [SUGGESTIONS](/topics/46433) topic of this script is a good place to ask if one could be put in or send me a PM and I'd be happy to consider it.

If you find any issues in my wiki fork by all means please let me know. I do my best to keep up to date but sometimes I do fall behind. Some pages aren't present yet but will be in due time. :)

69307.359673

ca2a91761ea404e535bc9ead696a0aaea0483d1b