105402.387043 uso - Dashboard Quick Menu [0.1.4](/scripts/diff/105402/387043)
- change menu- class to usermenu- with an implied id attribute of usermenu. Using the dash/hyphen (-) after the className pretty much imo indicates a "global wildcard"... so anything could happen.

More roughed logical reasoning behind this is based off of this...

className login_status is in use and should be isolated purely to USO core menu entries... so class of login_status- shouldn't be used. This is another reason for the hyphen at the end of the classname to avoid future conflicts with any luck. :)

105402.387043

SHA1:b8faeece89dfc0bc1416012bd11c02590835393b