68219.400347 uso - installWith [0.18.1](/scripts/diff/68219/400347)
- Add basic check for meta.js @version versus user.js @version comparison... this will now show a black install button with white text. The tooltip will show you what the respective @version values are... if you have "Skip embedded updater scan" disabled or a script doesn't have @version specified you will never see this Advisory.

Calling this "Security Advisory: ERROR, description follows" for now... you can still install a script that has this advisory however ANY UPDATE MECHANISM will be out of sync and you will NOT BE NOTIFIED of updates until [this userscripts.org caching issue](/topics/82159) is resolved. installWith WILL ABORT on this kind of ERROR,

What happens is the meta.js is current however the user.js is NOT... so if a script doesn't have @version in it there is currently no easy way to compare versions... as implied user.js is only parsed if embedded updater checking is enabled.  Since USO picks the first @version key for multiple @version keys this comparison is currently geared toward that.

- Removed useless @exclude rules

68219.400347

ac31188903a37578485854094f0d60a295f5c591