68219.403728 uso - installWith [1.0.0](/scripts/diff/68219/403728)
- Bumping @version to 1.0.0
- Use GM_setStyle lib to clean up some CSS
- Retooled entire code base and logic... this is step 1 in cleaning things up

Please note if there is a reference to a script homepage AND using an @require updater this is considered an embedded updater now... So previously colored Install buttons will be grey.

Also throbber is now geared towards Firefox 5.x+... if you need 4.x and below support then you will need to modify the CSS. I will post this later.

68219.403728

2d311c2baae549787dd95124d704b30145291cd2