69307.528942 uso - Count Issues [0.20.5](/scripts/diff/69307/528942)
- NEW FEATURE ... the "load source url bar"
This feature is in direct response to [this topic](/topics/118442) as I find it rather cumbersome to goto a 3rd party site which may or may not be online and why leave USO?! ;)

So the url mentioned in that topic for the dropbox... just paste the url into the load source url bar and tap enter/return.

This feature is still new and very roughed in but mostly works as expected. It also utilizes some newer HTML5 tags/methods. With a few CSS hacks I think that I have the combobox performing visually the way for this implementation, however I currently don't have a running Mac so I have no way of confirming the styling yet. Currently tested under Windows XP and Linux.

This feature will be improved upon but I thought we could all use it about now.

69307.528942

a4f3bc6eae8092165ee31dc17f201cebbfea3367