68219.813847 uso - installWith [2.0.2.0](/scripts/diff/68219/813847)
- NEW FEATURE (well old actually ;)

Exposed and created the user Advisor UI. Now you too can add your own advisories from the available tiers for yourself (e.g. block a User Script or a ScriptWright) as well as future sharing them in the Spam and Malware topic. A sample JSON file is included in the Advisor. This is in its infancy on the UI but it will improve with time. :)

- Bug fixed install button not showing on script home pages when Always hide potentially unwanted scripts was ticked.
- Improved visibility of script name on deleted scripts if showing
- Definitions update

68219.813847

7208091850fc8db1789e1c078a2b82ede53ca7bc